+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Prevent the mixture of the content for two pages that outputs to the same file [#170].
 - Consistency of `page.data.url` after change any `page.dest` value.
 - Bug detecting static files changes on update.
+- Improved the live-reload script for images and stylesheets.
 
 ## [1.5.1] - 2022-02-01
 ### Added